// playgrounds/schema/src/schemas/button.ts
// Generated by `localess schema pull`. Do not edit — the next pull overwrites this file.
import { defineField, defineSchema } from '@localess/schema';
import { ButtonType } from './button-type';

export const Button = defineSchema({
  id: 'Button',
  type: 'NODE',
  displayName: 'Button',
  description: 'A button',
  labels: ['button'],
  previewField: 'label',
  fields: [
    defineField({
      name: 'label',
      kind: 'TEXT',
      displayName: 'Label',
      required: true,
      translatable: true,
      description: 'Text that will appear inside the button',
      defaultValue: 'CTA',
      minLength: 3,
      maxLength: 30,
    }),
    defineField({
      name: 'type',
      kind: 'OPTION',
      displayName: 'Type',
      required: true,
      source: ButtonType,
    }),
  ],
});