// playgrounds/schema/src/schemas/page.ts
// Generated by `localess schema pull`. Do not edit — the next pull overwrites this file.
import { defineField, defineSchema } from '@localess/schema';
import { Button } from './button';

export const Page = defineSchema({
  id: 'Page',
  type: 'ROOT',
  displayName: 'Page',
  fields: [
    defineField({
      name: 'title',
      kind: 'TEXT',
      displayName: 'Title',
      translatable: true,
    }),
    defineField({
      name: 'description',
      kind: 'TEXTAREA',
      displayName: 'Description',
      translatable: true,
    }),
    defineField({
      name: 'buttons',
      kind: 'SCHEMAS',
      displayName: 'Buttons',
      schemas: [Button],
    }),
    defineField({
      name: 'content',
      kind: 'RICH_TEXT',
      displayName: 'Content',
      translatable: true,
      description: 'Rich Text example',
    }),
  ],
});